Compute National Accounts coverage ratios

Compute the National Accounts coverage ratios from the microdata estimates and items in National Accounts.

Usage

compute_coverage_ratios(
  df_estimates_microdata,
  na_table,
  formulas_ratios =
    as.list(set_names(names(nationalaccountslis::lis_dashboard_microdata_formulas)))
)

Arguments

df_estimates_microdata

A dataframe with the National Accounts estimates from the microdata. Must contain the ccyyd column and one column per variable, with column names matching the formulas in formulas_microdata. E.g. as returned by compute_estimates_lis_microdata().

na_table

A dataframe with the National Accounts data. E.g. as returned by produce_oecd_na_table(). Must contain the following columns: country, year and one column per variable. E.g. data.frame(country = c("ITA", "USA"), year = c(1985, 1985), D11R = c(100, 200), D12R = c(300, 400))

formulas_ratios

A list with strings containing formulas to produce the National Accounts ratios.

Value

A dataframe with the National Accounts ratios.